feat: move configuration pages to feature modules
Move starlark, manifests, mappings, and mcp-config pages from
pages/ to their respective feature module directories:

- pages/starlark/* → features/sagas/pages/
- pages/manifests/* → features/manifests/pages/
- pages/mappings/* → features/mappings/pages/
- pages/mcp-config/* → features/mcp-config/pages/

Update App.tsx imports to reference new locations. Add page
re-exports to each feature module barrel index.
