+++ frontend/src/App.tsx
@@ -19,10 +19,10 @@ import { PartiesPage, PartyDetailPage } from '@/features/parties'
 import { TenantsPage, TenantDetailPage } from '@/features/tenants'
 import { ReconciliationPage, ReconciliationDetailPage } from '@/features/reconciliation'
 import { AuditLogPage } from '@/pages/audit'
-import { StarlarkConfigPage } from '@/pages/starlark/index'
-import { StarlarkDetailPage } from '@/pages/starlark/detail'
-import { MappingsPage } from '@/pages/mappings'
-import { MappingDetailPage } from '@/pages/mappings/[mappingId]'
+import { StarlarkConfigPage } from '@/features/sagas/pages/index'
+import { StarlarkDetailPage } from '@/features/sagas/pages/detail'
+import { MappingsPage } from '@/features/mappings/pages'
+import { MappingDetailPage } from '@/features/mappings/pages/[mappingId]'
 import { ReferenceDataHubPage } from '@/pages/reference-data'
 import { InstrumentsPage } from '@/pages/reference-data/instruments'
 import { AccountTypesPage } from '@/pages/reference-data/account-types'
@@ -33,8 +33,8 @@ import { MarketDataPage } from '@/pages/market-data'
 import { DatasetDetailPage } from '@/pages/market-data/[datasetCode]'
 import { ForecastingPage } from '@/pages/forecasting'
 import { DashboardPage } from '@/pages/dashboard'
-import { ManifestsPage } from '@/pages/manifests'
-import { McpConfigPage } from '@/pages/mcp-config'
+import { ManifestsPage } from '@/features/manifests/pages'
+import { McpConfigPage } from '@/features/mcp-config/pages'
 import { TransactionsPage } from '@/pages/transactions'
 
 // Placeholder page components - replaced as each page task is implemented
