+++ frontend/src/features/manifests/index.ts
@@ -1,1 +1,2 @@
 // Public API for manifests module
+export { ManifestsPage } from './pages/index'
