+++ frontend/src/features/manifests/pages/manifest-apply-dialog.test.tsx
