+++ frontend/src/features/manifests/pages/manifest-apply-dialog.tsx
