+++ frontend/src/features/manifests/pages/manifest-current-view.test.tsx
