+++ frontend/src/features/manifests/pages/manifest-current-view.tsx
