+++ frontend/src/features/mappings/index.ts
@@ -1,1 +1,3 @@
 // Public API for mappings module
+export { MappingsPage } from './pages/index'
+export { MappingDetailPage } from './pages/[mappingId]'
