+++ frontend/src/features/mappings/pages/[mappingId].tsx
