+++ frontend/src/features/mcp-config/index.ts
@@ -1,1 +1,2 @@
 // Public API for mcp-config module
+export { McpConfigPage } from './pages/index'
