+++ frontend/src/features/mcp-config/pages/mcp-tools-section.tsx
