+++ frontend/src/features/sagas/index.ts
@@ -1,1 +1,4 @@
 // Public API for sagas module
+export { StarlarkConfigPage } from './pages/index'
+export { StarlarkDetailPage } from './pages/detail'
+export { CreateSagaDraftDialog } from './pages/create-saga-draft-dialog'
